@@ -1,11 +1,11 @@
 name: app_version_update
 description: An easy and quick way to check if the local app is updated with the same version in their respective stores (Play Store / Apple Store ).
 # author: Kauê Tomaz Murakami<kauetmurakami@gmail.com>
-version: 3.2.0
+version: 4.0.0
 homepage: https://github.com/kauemurakami/app_version_update
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: ">=2.18.0 <4.1.0"
   flutter: ">=1.17.0"
 
 dependencies:
@@ -13,7 +13,7 @@ dependencies:
     sdk: flutter
 
   http: 
-  package_info_plus: 
+  package_info_plus: ^4.0.0
   url_launcher:
   # http: ^0.13.5
   # package_info_plus: ^1.4.3+1
